Refactor finance view rendering modules
Run node --test tests/ui/cards/financeLedger.test.js tests/ui/cards/financeObligations.test.js tests/ui/cards/financeOpportunities.test.js tests/ui/cards/financePulse.test.js
