+++ src/ui/views/browser/apps/finance.js
@@ -2,200 +2,11 @@ import { formatHours, formatMoney } from '../../../../core/helpers.js';
 import { getPageByType } from './pageLookup.js';
 import renderFinanceHeader from './finance/header.js';
 import renderFinanceLedger, { createBankSection } from './finance/ledger.js';
-import { formatCurrency, formatSignedCurrency } from '../utils/financeFormatting.js';
-
-function renderFinanceHistory(history = []) {
-  const { section, body } = createBankSection(
-    'Cashflow History',
-    'Seven-day snapshots captured at day end.'
-  );
-
-  const entries = Array.isArray(history) ? history.slice(0, 7) : [];
-  if (!entries.length) {
-    const empty = document.createElement('p');
-    empty.className = 'bankapp-empty';
-    empty.textContent = 'Wrap a full day to start the rolling history.';
-    body.appendChild(empty);
-    return section;
-  }
-
-  const list = document.createElement('ol');
-  list.className = 'bankapp-history';
-
-  entries.forEach(entry => {
-    const item = document.createElement('li');
-    item.className = 'bankapp-history__item';
-    if (entry?.tone) {
-      item.dataset.tone = entry.tone;
-    }
-
-    const header = document.createElement('div');
-    header.className = 'bankapp-history__header';
-    const label = document.createElement('span');
-    label.className = 'bankapp-history__label';
-    label.textContent = entry?.label || 'Day';
-    header.appendChild(label);
-
-    if (Number.isFinite(entry?.recordedAt)) {
-      const time = document.createElement('time');
-      time.className = 'bankapp-history__time';
-      const stamp = new Date(entry.recordedAt);
-      time.dateTime = stamp.toISOString();
-      time.textContent = stamp.toLocaleString(undefined, {
-        month: 'short',
-        day: 'numeric',
-        hour: 'numeric',
-        minute: '2-digit'
-      });
-      header.appendChild(time);
-    }
-
-    const totals = document.createElement('div');
-    totals.className = 'bankapp-history__totals';
-    const net = document.createElement('span');
-    net.className = 'bankapp-history__net';
-    net.textContent = formatSignedCurrency(entry?.totals?.net || 0);
-    totals.appendChild(net);
-
-    const split = document.createElement('span');
-    split.className = 'bankapp-history__split';
-    const income = Number(entry?.totals?.income || 0);
-    const spend = Number(entry?.totals?.spend || 0);
-    split.textContent = `${formatCurrency(income)} • ${formatCurrency(spend > 0 ? -spend : 0)}`;
-    totals.appendChild(split);
-
-    const highlights = document.createElement('p');
-    highlights.className = 'bankapp-history__highlights';
-    const incomeTop = entry?.ledger?.payouts?.[0];
-    const spendTop = entry?.ledger?.costs?.[0];
-    const details = [];
-    if (incomeTop) {
-      details.push(
-        `${incomeTop.label || 'Income'} ${formatSignedCurrency(incomeTop.amount || 0)}`
-      );
-    }
-    if (spendTop) {
-      details.push(
-        `${spendTop.label || 'Spend'} ${formatSignedCurrency(
-          spendTop.amount ? -spendTop.amount : 0
-        )}`
-      );
-    }
-    highlights.textContent = details.length
-      ? details.join(' • ')
-      : 'Cashflow steady without standout spikes.';
-
-    item.append(header, totals, highlights);
-    list.appendChild(item);
-  });
-
-  body.appendChild(list);
-  return section;
-}
-
-function renderFinanceActivity(entries = []) {
-  const { section, body } = createBankSection(
-    'Recent Activity Log',
-    'Latest timeline pulled from the shared activity feed.'
-  );
-
-  const activity = Array.isArray(entries) ? entries.slice(0, 10) : [];
-  if (!activity.length) {
-    const empty = document.createElement('p');
-    empty.className = 'bankapp-empty';
-    empty.textContent = 'No recent entries yet—start hustling to fill the log!';
-    body.appendChild(empty);
-    return section;
-  }
-
-  const list = document.createElement('ul');
-  list.className = 'bankapp-activity';
-
-  activity.forEach(entry => {
-    const item = document.createElement('li');
-    item.className = 'bankapp-activity__item';
-    if (entry?.tone) {
-      item.dataset.tone = entry.tone;
-    }
-
-    const message = document.createElement('span');
-    message.className = 'bankapp-activity__message';
-    message.textContent = entry?.message || '';
-    item.appendChild(message);
-
-    if (Number.isFinite(entry?.timestamp)) {
-      const time = document.createElement('time');
-      const stamp = new Date(entry.timestamp);
-      time.className = 'bankapp-activity__time';
-      time.dateTime = stamp.toISOString();
-      time.textContent = stamp.toLocaleTimeString(undefined, { hour: 'numeric', minute: '2-digit' });
-      item.appendChild(time);
-    }
-
-    list.appendChild(item);
-  });
-
-  body.appendChild(list);
-  return section;
-}
-
-function renderFinanceObligations(model = {}) {
-  const { section, body } = createBankSection('Pending & Upcoming Obligations', 'Keep upkeep, payroll, and tuition funded.');
-  const entries = Array.isArray(model.entries) ? model.entries : [];
-
-  if (!entries.length) {
-    const empty = document.createElement('p');
-    empty.className = 'bankapp-empty';
-    empty.textContent = 'No obligations queued. Everything is funded!';
-    body.appendChild(empty);
-    return section;
-  }
-
-  const grid = document.createElement('div');
-  grid.className = 'bankapp-obligations';
-
-  entries.forEach(entry => {
-    const card = document.createElement('article');
-    card.className = 'bankapp-card bankapp-card--obligation';
-    const header = document.createElement('div');
-    header.className = 'bankapp-card__header';
-    const title = document.createElement('h3');
-    title.textContent = entry.label || 'Obligation';
-    const amount = document.createElement('span');
-    amount.className = 'bankapp-card__amount';
-    amount.textContent = formatCurrency(entry.amount || 0);
-    header.append(title, amount);
-    card.appendChild(header);
-
-    if (entry.note) {
-      const note = document.createElement('p');
-      note.className = 'bankapp-card__note';
-      note.textContent = entry.note;
-      card.appendChild(note);
-    }
-
-    if (Array.isArray(entry.items) && entry.items.length) {
-      const list = document.createElement('ul');
-      list.className = 'bankapp-card__list';
-      entry.items.forEach(item => {
-        const row = document.createElement('li');
-        row.className = 'bankapp-card__list-item';
-        const name = document.createElement('span');
-        name.textContent = item.label || 'Entry';
-        const value = document.createElement('span');
-        value.textContent = formatCurrency(item.amount || 0);
-        row.append(name, value);
-        list.appendChild(row);
-      });
-      card.appendChild(list);
-    }
-
-    grid.appendChild(card);
-  });
-
-  body.appendChild(grid);
-  return section;
-}
+import renderFinanceActivity from './finance/activity.js';
+import renderFinanceEducation from './finance/education.js';
+import renderFinanceHistory from './finance/history.js';
+import renderFinanceObligations from './finance/obligations.js';
+import { formatCurrency } from '../utils/financeFormatting.js';
 
 function renderFinancePendingIncome(entries = []) {
   const { section, body } = createBankSection('In-Flight Earnings', 'Assets with payouts pending the next day rollover.');
@@ -400,57 +211,6 @@ function renderFinanceOpportunities(model = {}) {
   return section;
 }
 
-function renderFinanceEducation(entries = []) {
-  const { section, body } = createBankSection('Education Investments', 'Courses in progress with tuition already committed.');
-  const list = Array.isArray(entries) ? entries : [];
-
-  if (!list.length) {
-    const empty = document.createElement('p');
-    empty.className = 'bankapp-empty';
-    empty.textContent = 'No active courses. Enroll in a study track to plan tuition.';
-    body.appendChild(empty);
-    return section;
-  }
-
-  const grid = document.createElement('div');
-  grid.className = 'bankapp-education';
-
-  list.forEach(entry => {
-    const card = document.createElement('article');
-    card.className = 'bankapp-card bankapp-card--education';
-    const header = document.createElement('div');
-    header.className = 'bankapp-card__header';
-    const title = document.createElement('h3');
-    title.textContent = entry.name || 'Course';
-    const tuition = document.createElement('span');
-    tuition.className = 'bankapp-card__amount';
-    tuition.textContent = entry.tuition > 0 ? formatCurrency(entry.tuition) : 'Free';
-    header.append(title, tuition);
-    card.appendChild(header);
-
-    const note = document.createElement('p');
-    note.className = 'bankapp-card__note';
-    note.textContent = `${entry.remainingDays} day${entry.remainingDays === 1 ? '' : 's'} left • ${formatHours(entry.hoursPerDay || 0)}/day`;
-    card.appendChild(note);
-
-    if (entry.bonus) {
-      const bonus = document.createElement('p');
-      bonus.className = 'bankapp-card__note bankapp-card__note--muted';
-      bonus.textContent = entry.bonus;
-      card.appendChild(bonus);
-    }
-
-    const status = document.createElement('p');
-    status.className = 'bankapp-card__status';
-    status.textContent = entry.studiedToday ? 'Today’s study scheduled' : 'Waiting for today’s study slot';
-    card.appendChild(status);
-
-    grid.appendChild(card);
-  });
-
-  body.appendChild(grid);
-  return section;
-}
 
 export default function renderFinance(context = {}, registries = {}, models = {}) {
   const page = getPageByType('finance');
